A simple blackjack game
Blackjack is a game played against a dealer, with each of player and dealer beginning with 2 cards. The goal is to get closer to 21 than the dealer by drawing more cards, but with the catch being that if the player goes over 21, the player immediately loses. The dealer will draw until their hand is worth at least 17, and the player can decide when to draw, and when to hold. Face cards are worth 10, and (in our version) Aces are worth 11. Ties go to the dealer.